Osa4 demot (#19)
* Aloitettu käyttäjänhallinnan toteutus

* Osa 4 c done

* Tokenin käyttö done

* Osan 4 demot valmiit

* Copypasted tulilla tehtäviä varten

* Osa 4 T 4.16 done

* Osa 4 T 4.17 done

* Osa 4 T 4.19 done

* Osa 4 T 4.20 done

* Osa 4 valmis

+++ osa4/controllers/blogs.js
@@ -1,21 +1,30 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({})
+        .populate('user', { username: 1, name: 1, id: 1 })
+
     response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-    const blog = new Blog(request.body)
-    blog.likes === undefined
-        ? blog.likes = 0
-        : blog.likes
+    const user = await validateAndGetUser(request, response, next)
+    if (user === undefined) {
+        return
+    }
 
     try {
+        const blog = createBlog(request.body, user._id)
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog)
+        await user.save()
+
         response.status(201)
-        response.json(savedBlog)
+        response.json(savedBlog.toJSON())
     } catch (exception) {
         next(exception)
     }
@@ -38,12 +47,45 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+    const user = await validateAndGetUser(request, response, next)
+    if (user === undefined) {
+        return
+    }
+
     try {
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
+        const blog = await Blog.findById(request.params.id)
+
+        if (blog.user.toString() === user.id.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        }
     } catch (exception) {
         next(exception)
     }
 })
+
+// Utility functions:
+const createBlog = (body, userId) => {
+    return new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes || 0,
+        user: userId,
+    })
+}
+
+const validateAndGetUser = async (request, response, next) => {
+    try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+
+        return await User.findById(decodedToken.id)
+    } catch (error) {
+        next(error)
+    }
+}
 
 module.exports = blogsRouter
